requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
Flask-Assets==0.11
Flask-Migrate==1.3.1
Flask-RESTful==0.3.4
Flask-SQLAlchemy==2.0
psycopg2==2.6.1
SQLAlchemy==1.0.5
SQLAlchemy-Utils==0.30.5
PyJWT==1.4.0
uWSGI==2.0.11.1
requests==2.8.1
docopt==0.6.2
marshmallow==1.2.6